Menu UI Design started
Also, a draft of the interactive tutorial, as well as a "tutorial.txt" file containing the dialogs of the interactive tutorial have been prepared for future development. As it being a tedious work and not an absolute necessity compared to designing the menus, it is left as it is for now.